Add eventHandler to Cell class
