+++ src/main/java/ttt/Cell.java
@@ -1,5 +1,7 @@
 package ttt;
 
+import javafx.event.EventHandler;
+import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.*;
 import javafx.scene.paint.Color;
 import javafx.scene.text.Text;
@@ -15,6 +17,11 @@ public class Cell extends StackPane{
         this.setBorder(new Border(new BorderStroke(Color.BLACK, BorderStrokeStyle.SOLID, CornerRadii.EMPTY, new BorderWidths(2))));
         GridPane.setRowIndex(this,(index-1)%3);
         GridPane.setColumnIndex(this,(index-1)/3);
+        this.setOnMousePressed(new EventHandler<MouseEvent>() {
+            public void handle(MouseEvent me) {
+                handleClick(me);
+            }
+        });
     }
 
     public void setMarker(String marker) {
@@ -23,5 +30,13 @@ public class Cell extends StackPane{
 
     public int getIndex() {
         return this.index;
+    }
+
+    public static void handleClick(MouseEvent event) {
+        StackPane source = (StackPane) event.getSource();
+        GridPane parent = (GridPane) source.getParent();
+        int row = parent.getRowIndex(source);
+        int column = parent.getColumnIndex(source);
+        System.out.println(row + ", " + column);
     }
 }
